Move Literal from typing_extensions
