+++ typing_extensions/src_py2/typing_extensions.py
@@ -4,7 +4,7 @@ import typing
 from typing import (
     ClassVar, Type, Generic, Callable, GenericMeta, TypingMeta,
     Counter, DefaultDict, Deque, TypeVar, Tuple, Final, final,
-    NewType, overload, Text, TYPE_CHECKING,
+    NewType, overload, Text, TYPE_CHECKING, Literal,
     # We use internal typing helpers here, but this significantly reduces
     # code duplication. (Also this is only until Protocol is in typing.)
     _generic_new, _type_vars, _next_in_mro, _tp_cache, _type_check,
@@ -113,68 +113,6 @@ def _gorg(cls):
 
 def IntVar(name):
     return TypeVar(name)
-
-
-class _LiteralMeta(TypingMeta):
-    """Metaclass for _Literal"""
-
-    def __new__(cls, name, bases, namespace):
-        cls.assert_no_subclassing(bases)
-        self = super(_LiteralMeta, cls).__new__(cls, name, bases, namespace)
-        return self
-
-
-class _Literal(typing._FinalTypingBase):
-    """A type that can be used to indicate to type checkers that the
-    corresponding value has a value literally equivalent to the
-    provided parameter. For example:
-
-        var: Literal[4] = 4
-    
-    The type checker understands that 'var' is literally equal to the
-    value 4 and no other value.
-
-    Literal[...] cannot be subclassed. There is no runtime checking
-    verifying that the parameter is actually a value instead of a type.
-    """
-
-    __metaclass__ = _LiteralMeta
-    __slots__ = ('__values__',)
-
-    def __init__(self, values=None, **kwds):
-        self.__values__ = values
-
-    def __getitem__(self, item):
-        cls = type(self)
-        if self.__values__ is None:
-            if not isinstance(item, tuple):
-                item = (item,)
-            return cls(values=item,
-                       _root=True)
-        raise TypeError('{} cannot be further subscripted'
-                        .format(cls.__name__[1:]))
-
-    def _eval_type(self, globalns, localns):
-        return self
-
-    def __repr__(self):
-        r = super(_Literal, self).__repr__()
-        if self.__values__ is not None:
-            r += '[{}]'.format(', '.join(map(typing._type_repr, self.__values__)))
-        return r
-
-    def __hash__(self):
-        return hash((type(self).__name__, self.__values__))
-
-    def __eq__(self, other):
-        if not isinstance(other, _Literal):
-            return NotImplemented
-        if self.__values__ is not None:
-            return self.__values__ == other.__values__
-        return self is other
-
-
-Literal = _Literal(_root=True)
 
 
 class _ProtocolMeta(GenericMeta):
